refactor(nav-drawer): move logic to function
