+++ src/scripts/components/stories/pilot-story/components/utils/nav-drawer.ts
@@ -1,4 +1,5 @@
 import {RefObject} from 'react';
+import {dataIsTitleInView} from '../../config/main';
 
 /**
  * Calculates the initial snap point for the mobile drawer based on the height of the handler element and the window height.
@@ -19,20 +20,37 @@ export const updateIndicatorPosition = (
   indicationRef: RefObject<HTMLElement>,
   length: number,
   selectedChapterIndex: number,
-  progress: number
+  progress: number,
+  gap: number
 ) => {
-  const progressIndicatorHeight = indicationRef.current?.clientHeight;
+  if (!indicationRef.current) {
+    return;
+  }
+  indicationRef.current.setAttribute(dataIsTitleInView, 'false');
+
+  const progressIndicatorHeight = indicationRef.current.clientHeight;
+
+  // Get the circle radius from the CSS custom property
+  // This is used to calculate the offset of the indicator
+  // We want the indicator to start below the circle indication for the selected chapter intro
+  const circleRadius = Number(
+    window
+      .getComputedStyle(indicationRef.current)
+      .getPropertyValue('--circle-radius')
+      .replace('px', '')
+  );
 
   if (progressIndicatorHeight && length) {
     const indicatorYOffsetInPx =
-      (progressIndicatorHeight / (length + 1)) * selectedChapterIndex +
-      (progressIndicatorHeight / (length + 1)) * progress;
+      (progressIndicatorHeight / length) * selectedChapterIndex +
+      (progressIndicatorHeight / length) * progress +
+      (selectedChapterIndex * gap) / circleRadius;
 
     const indicatorYOffsetInPercent = `${
       (indicatorYOffsetInPx / progressIndicatorHeight) * 100
     }%`;
 
-    indicationRef.current?.style.setProperty(
+    indicationRef.current.style.setProperty(
       '--indicator-y-offset',
       indicatorYOffsetInPercent
     );
